Fixed rx handling
bsky docs suggest matching on bytes instead of strings, because the
index positions of facets are byte based.  But this kills match patterns
like \w which won't match umlauts anymore in a bytes sequence.

The solution is to match in utf8, then convert the prefix before the
match to bytes and use that length as start.  Same conversion with
finding the end.